moved flush
